(PC-25607) refacto(offer): made a component for about section to avoid complexity cognitive and seperate tests
